refactor: cambiado nombre de entidad de gestion_usuarios a users
