+++ backend/fisio_find/gestion_citas/exampledata.json
@@ -1,48 +1,48 @@
 [
     {
-        "model": "gestion_usuarios.specialization",
+        "model": "users.specialization",
         "pk": 1,
         "fields": {
             "name": "Deportiva"
         }
     },
     {
-        "model": "gestion_usuarios.specialization",
+        "model": "users.specialization",
         "pk": 2,
         "fields": {
             "name": "Ortopédica y Traumatológica"
         }
     },
     {
-        "model": "gestion_usuarios.specialization",
+        "model": "users.specialization",
         "pk": 3,
         "fields": {
             "name": "Geriátrica"
         }
     },
     {
-        "model": "gestion_usuarios.specialization",
+        "model": "users.specialization",
         "pk": 4,
         "fields": {
             "name": "Pediátrica"
         }
     },
     {
-        "model": "gestion_usuarios.specialization",
+        "model": "users.specialization",
         "pk": 5,
         "fields": {
             "name": "Neurológica"
         }
     },
     {
-        "model": "gestion_usuarios.specialization",
+        "model": "users.specialization",
         "pk": 6,
         "fields": {
             "name": "Obstetricia y suelo pélvico"
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 1,
         "fields": {
             "username": "jorgito",
@@ -57,7 +57,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.physiotherapist",
+        "model": "users.physiotherapist",
         "pk": 1,
         "fields": {
             "user": 1,
@@ -137,7 +137,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 2,
         "fields": {
             "username": "abel",
@@ -152,7 +152,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.physiotherapist",
+        "model": "users.physiotherapist",
         "pk": 2,
         "fields": {
             "user": 2,
@@ -171,7 +171,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 3,
         "fields": {
             "username": "lupe",
@@ -186,7 +186,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.physiotherapist",
+        "model": "users.physiotherapist",
         "pk": 3,
         "fields": {
             "user": 3,
@@ -205,7 +205,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 4,
         "fields": {
             "username": "guille",
@@ -220,7 +220,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.physiotherapist",
+        "model": "users.physiotherapist",
         "pk": 4,
         "fields": {
             "user": 4,
@@ -238,7 +238,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 5,
         "fields": {
             "username": "lupita",
@@ -253,7 +253,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.physiotherapist",
+        "model": "users.physiotherapist",
         "pk": 5,
         "fields": {
             "user": 5,
@@ -272,7 +272,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 6,
         "fields": {
             "username": "patient1",
@@ -287,7 +287,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.patient",
+        "model": "users.patient",
         "pk": 1,
         "fields": {
             "user": 6,
@@ -296,7 +296,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 7,
         "fields": {
             "username": "patient2",
@@ -309,7 +309,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.patient",
+        "model": "users.patient",
         "pk": 2,
         "fields": {
             "user": 7,
@@ -318,7 +318,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 8,
         "fields": {
             "username": "patient3",
@@ -331,7 +331,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.patient",
+        "model": "users.patient",
         "pk": 3,
         "fields": {
             "user": 8,
@@ -340,7 +340,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 9,
         "fields": {
             "username": "patient4",
@@ -353,7 +353,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.patient",
+        "model": "users.patient",
         "pk": 4,
         "fields": {
             "user": 9,
@@ -362,7 +362,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.appuser",
+        "model": "users.appuser",
         "pk": 10,
         "fields": {
             "username": "patient5",
@@ -375,7 +375,7 @@
         }
     },
     {
-        "model": "gestion_usuarios.patient",
+        "model": "users.patient",
         "pk": 5,
         "fields": {
             "user": 10,
