+++ backend/fisio_find/sesion_invitado/models.py
@@ -1,4 +1,4 @@
 from django.db import models
-from gestion_usuarios.models import Admin
+from users.models import Admin
 
 # Create your models here.
