+++ backend/fisio_find/users/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class GestionUsuariosConfig(AppConfig):
+class UsersConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'gestion_usuarios'
+    name = 'users'
